Invalid input handled correctly for Regions.js

+++ server/src/routes/Regions.js
@@ -9,22 +9,10 @@ type Response = express$Response;
 
 module.exports = {
   getAllRegionsInCounty: function(req: Request, res: Response) {
-    return Region.findAll({ where: { county_id: Number(req.params.county_id) } }).then(regions =>
-      regions ? res.send(regions) : res.sendStatus(404)
+    if (!req.params || isNaN(Number(req.params.county_id))) return res.sendStatus(400);
+    return Region.findAll({ where: { county_id: Number(req.params.county_id) } }).then(
+      regions => (regions ? res.send(regions) : res.sendStatus(404))
     );
-  },
-  getRegionName: function(req: Request, res: Response) {
-    return Region.findOne({ where: { region_id: res.body.region_id }, attributes: ['name'] }).then(name =>
-      name ? res.send(name) : res.sendStatus(404)
-    );
-  },
-  getOneRegionByNameAndCounty: async function(req: Request, res: Response) {
-    let c_id = await County.getOneCountyByName(req, res);
-    let countyId = c_id ? c_id : res.sendStatus(404);
-    return Region.findOne({
-      where: { name: req.params.region_name, county_id: Number(countyId.county_id) },
-      attributes: ['region_id']
-    }).then(regions => (regions ? regions : res.sendStatus(404)));
   },
   getAllRegions: function(req: Request, res: Response) {
     return Region.findAll().then(regions => res.send(regions));
@@ -47,15 +35,16 @@ module.exports = {
     }).then(regions => (regions ? res.send(regions) : res.sendStatus(404)));
   },
   getRegion: function(req: Request, res: Response) {
-    return Region.findOne({ where: { region_id: Number(req.params.region_id) } }).then(region =>
-      region ? res.send(region) : res.sendStatus(404)
+    if (!req.params || isNaN(Number(req.params.region_id))) return res.sendStatus(400);
+    return Region.findOne({ where: { region_id: Number(req.params.region_id) } }).then(
+      region => (region ? res.send(region) : res.sendStatus(404))
     );
   },
   updateRegion: function(req: Request, res: Response) {
-    let region_id = Number(req.params.region_id);
     if (
       !req.body ||
-      typeof region_id !== 'number' ||
+      !req.params ||
+      isNaN(Number(req.params.region_id)) ||
       typeof req.body.name !== 'string' ||
       typeof req.body.lat !== 'number' ||
       typeof req.body.lon !== 'number' ||
@@ -71,10 +60,12 @@ module.exports = {
         lon: req.body.lon,
         county_id: req.body.county_id
       },
-      { where: { region_id: region_id } }
+      { where: { region_id: Number(req.params.region_id) } }
     ).then(regions => (regions ? res.send(regions) : res.sendStatus(404)));
   },
   delRegion: function(req: Request, res: Response) {
+    if (!req.params || isNaN(Number(req.params.region_id))) return res.sendStatus(400);
+
     return Region.destroy({ where: { region_id: Number(req.params.region_id) } })
       .then(regions => (regions ? res.send() : res.status(500).send()))
       .catch(err => {
@@ -86,12 +77,11 @@ module.exports = {
   },
 
   getRegionStaff: function(req: Request, res: Response) {
-    if (!req.params || typeof Number(req.params.region_id) != 'number') return res.sendStatus(400);
+    if (!req.params || isNaN(Number(req.params.region_id))) return res.sendStatus(400);
 
-    const region_id = Number(req.params.region_id);
     User.findAll({
       where: {
-        region_id: region_id,
+        region_id: Number(req.params.region_id),
         role_id: 2
       }
     })
